% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{addBKVExprGroups_to_Seurat}
\alias{addBKVExprGroups_to_Seurat}
\title{addBKVExprGroups_to_Seurat}
\usage{
addBKVExprGroups_to_Seurat(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
A Seurat object with 4 new columns in @meta.data
}
\description{
This function adds 4 columns to slot @meta.data of a Seurat object. They are 1) VP1_group, 2) VP1_group_genotype, 3) LTAg_group, and 4) LTAg_group_genotype.
}
\examples{
\dontrun{
seurat = addBKVExprGroups_to_Seurat(seurat)
}
}
\keyword{Seurat}
