% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{addBKVFracExpression_to_Seurat}
\alias{addBKVFracExpression_to_Seurat}
\title{addBKVFracExpression_to_Seurat}
\usage{
addBKVFracExpression_to_Seurat(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
A Seurat object with 7 new columns in @meta.data
}
\description{
This function adds seven columns to slot @meta.data of a Seurat object. They are 1) Total_mRNA, 2) counts_VP1, 3) counts_LTAg, 4) counts_BKV, 5) frac_VP1, 6) frac_LTAg, and 7) frac_BKV. 'Total_mRNA' is the sum of the raw UMI counts in each cell. The 'counts' are sum of raw counts for VP1 or LTAg in each cell. 'Frac' is the fraction of VP1 (or LTAg, BKV) counts in each cell using 'Total_mRNA'.
}
\examples{
seurat = addBKVFracExpression_to_Seurat(seurat)
}
\keyword{Seurat}
