% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{addUpperLowerBound_to_Seurat}
\alias{addUpperLowerBound_to_Seurat}
\title{addUpperLowerBound_to_Seurat}
\usage{
addUpperLowerBound_to_Seurat(seurat, pheno.use = "Total_mRNA")
}
\arguments{
\item{seurat}{A Seurat object}

\item{pheno.use}{The @meta.data column name to use (default = 'Total_mRNA')}
}
\value{
A Seurat object with 3 new columns in @meta.data
}
\description{
This function adds three columns to slot @meta.data of a Seurat object. They are 1) Filter_high_mRNA, 2) Filter_low_mRNA, and 3) Filter_ok and all three are logical (T or F) vectors. Function uses same calculation as Monocle documentation to identify cells with significantly low and high Total_mRNA counts.
}
\details{
From Monocle documentation: removed the cells with very low mRNA recovery or far more mRNA that the typical cell. Often, doublets or triplets have roughly twice the mRNA recovered as true single cells, so the latter filter is another means of excluding all but single cells from the analysis. Such filtering is handy if your protocol doesn't allow directly visualization of cell after they've been captured.
}
\examples{
\dontrun{
seurat = addUpperLowerBound_to_Seurat(seurat)
seurat = addUpperLowerBound_to_Seurat(seurat, pheno.use = 'nUMI')
}
}
\keyword{Seurat}
