% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{densityPlot_Seurat}
\alias{densityPlot_Seurat}
\title{densityPlot_Seurat}
\usage{
densityPlot_Seurat(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
Nothing
}
\description{
Outputs a Density plot for all cells. The data used will be the same metadata column name used in the addUpperLowerBound_to_Seurat() function call. Draws vertical lines to show the lower and upper bound cutoff.
}
\examples{
\dontrun{
densityPlot_Seurat(seurat)
}
}
\keyword{Seurat}
