% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindIdentLabel.R
\name{FindIdentLabel}
\alias{FindIdentLabel}
\title{FindIdentLabel}
\usage{
FindIdentLabel(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
The colname in metadata
}
\description{
Find identical label between ident and metadata
}
\examples{
\dontrun{
FindIdentLabel(seurat)
}
}
\references{
https://github.com/nyuhuyang/SeuratExtra/blob/master/R/Seurat4_functions.R
}
