% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixClusterFactorLevels.R
\name{FixClusterFactorLevels}
\alias{FixClusterFactorLevels}
\title{FixClusterFactorLevels}
\usage{
FixClusterFactorLevels(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\value{
A Seurat object
}
\description{
Relevel cluster factor levels so they are in numerical order
}
\examples{

FixClusterFactorLevels(pbmc_small)

}
