% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixFragmentPaths.R
\name{FixFragmentPaths}
\alias{FixFragmentPaths}
\title{FixFragmentPaths}
\usage{
FixFragmentPaths(
  seurat,
  cellranger_path,
  assay = "ATAC",
  sample_col = "orig.ident"
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cellranger_path}{Path to the cellranger folder that contains subfolders matching the sample names in the Seurat metadata `orig.ident`}

\item{assay}{The assay name for the ChromatinAssay in the Seurat object}

\item{sample_col}{The column name in Seurat metadata that contains the sample identifiers}
}
\value{
A Seurat object
}
\description{
Fix the paths to the atac_fragments.tsv.gz file in a Seurat object
}
\examples{
\dontrun{
FixFragmentPaths(seurat, cellranger_path = "../cellranger_count")
}

}
