% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{output_process_heatmaps}
\alias{output_process_heatmaps}
\title{output_process_heatmaps}
\usage{
output_process_heatmaps(seurat_align, markers, pdffile, ...)
}
\arguments{
\item{seurat_align}{A Seurat alignment object}

\item{markers}{Marker table from Seurat}

\item{pdffile}{PDF filename to save results}
}
\value{
Nothing
}
\description{
Wrapper around process_heatmap() that outputs all heatmaps to a PDF.
}
\examples{
output_process_heatmaps(aligned, markers, "processheatmaps.alignment.pdf")
}
\keyword{Seurat}
\keyword{heatmap}
\keyword{processheatmaps}
