% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small_v5}
\alias{pbmc_small_v5}
\title{PBMC Seurat object (version 5)}
\format{
Seurat object with 230 rows/genes and 80 columns/cells.
\describe{
  \item{nCount_RNA}{Number of UMI counts in the cell}
  \item{nFeature_RNA}{Number of genes detected in the cell}
  \item{RNA_snn_res.0.8}{Cluster identity using resolution 0.8}
  \item{RNA_snn_res.1}{Cluster identity using resolution 1}
  \item{letter.idents}{Sample id?}
  \item{groups}{Group id}
}
}
\source{
https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k
}
\usage{
pbmc_small_v5
}
\description{
Converted pbmc_small into a version 5 Seurat object.
}
\references{
Seurat package https://satijalab.org/seurat/
}
\keyword{datasets}
