% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.seurat_info.R
\name{print.seurat_info}
\alias{print.seurat_info}
\title{Print a seurat_info object}
\usage{
\method{print}{seurat_info}(x, ...)
}
\arguments{
\item{x}{A `seurat_info` object, as returned by [SeuratInfo()]}

\item{...}{Ignored}
}
\value{
`x`, invisibly
}
\description{
Formats and prints the report gathered by [SeuratInfo()]. All of the
formatting for that report lives here; `SeuratInfo()` only gathers the data.
}
