% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{process_heatmap}
\alias{process_heatmap}
\title{Process heatmap}
\usage{
process_heatmap(object, markers, process, colors = NULL, fontsizeRow = 4, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{markers}{Marker table from Seurat}

\item{process}{Process name (from gene annotations)}

\item{fontsizeRow}{Font size of row labels in Pheatmap}
}
\value{
List of 3 objects: 2 pheatmaps (unclustered and clustered) and scaled data
}
\description{
Builds an unclustered and a clustered scaled heatmap for genes in a Process from Marker tables.
}
\examples{
\dontrun{
process_heatmaps(seurat, markers, process = process, ...)
}
}
\keyword{Seurat}
\keyword{pheatmap}
