% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadWriteSeurat.R
\name{ReadSeurat}
\alias{ReadSeurat}
\title{ReadSeurat}
\usage{
ReadSeurat(path)
}
\arguments{
\item{path}{Path to the file (.rds or .qs2)}
}
\value{
A Seurat object
}
\description{
Read a Seurat object from an RDS or QS2 file.
}
\examples{

tmp <- tempfile(fileext = ".rds")
WriteSeurat(pbmc_small, tmp)
seurat <- ReadSeurat(tmp)
unlink(tmp)

}
