% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{relevel_groupgenotype_in_Seurat}
\alias{relevel_groupgenotype_in_Seurat}
\title{relevel_groupgenotype_in_Seurat}
\usage{
relevel_groupgenotype_in_Seurat(seurat, name, groups, bkvname)
}
\arguments{
\item{seurat}{A Seurat object}

\item{name}{The meta.data field that needs releveling}

\item{groups}{i.e. c("0%", "1%", "10%", "100%")}

\item{bkvname}{either of c("BKVm5", "LT")}
}
\value{
A Seurat object with releveled 'name' field in @meta.data
}
\description{
This function relevels the group_genotype meta_data field so that Mock comes before 'bkvname' and second, in order of the 'groups'.
}
\keyword{Seurat}
