% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{Self_scmapCluster}
\alias{Self_scmapCluster}
\title{Self_scmapCluster}
\usage{
Self_scmapCluster(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object (returned object from Seurat2SingleCellExperiment) with colData populated with 1 or more columns.}
}
\description{
From a SingleCellExperiment object, this function performs a scmap cluster analysis on itself. Useful for validating clusters.
}
\examples{
names = c("Detox", "DNAReplication" "Quiescent")
sce = Seurat2SingleCellExperiment(seurat, names) # 'seurat' is created through use of Seurat package
Self_scmapCluster(sce)
}
\keyword{scmap}
