% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sct.R
\name{Seurat2SingleCellExperiment}
\alias{Seurat2SingleCellExperiment}
\title{Seurat2SingleCellExperiment}
\usage{
Seurat2SingleCellExperiment(seurat, clusterlabels = NULL)
}
\arguments{
\item{seurat}{A Seurat object}

\item{clusterlabels}{Optional character vector that is same length as the number of clusters. For example, if three clusters (0, 1, 2) then clusterlabels might be c("a","b","c").}
}
\value{
A SingleCellExperiment object
}
\description{
This function creates a Bioconductor SingleCellExperiment from a Seurat object. Raw counts are extracted for the cells used in building the Seurat clusters. The raw counts are normalized by 'scater' package. If you get the following warning message you can ignore it (it comes from scater normalize function): In .local(object, ...) : using library sizes as size factors.
}
\examples{
Seurat2SingleCellExperiment(seurat, c(1,2,3))   # 'seurat' is created through use of Seurat package
}
\keyword{Bioconductor}
\keyword{Seurat}
\keyword{SingleCellExperiment}
