% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignacInfo.R
\name{SignacInfo}
\alias{SignacInfo}
\title{SignacInfo}
\usage{
SignacInfo(seurat, assay = NULL, show_annotation = FALSE)
}
\arguments{
\item{seurat}{A Seurat object}

\item{assay}{Assay to get information for (default is all ChromAssays)}

\item{show_annotation}{Show genome annotation information (default FALSE)}
}
\description{
Show information about the Signac ChromatinAssays within a Seurat object
}
\examples{
SignacInfo(multiome_small)
}
