% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateMetadataCols.R
\name{ValidateMetadataCols}
\alias{ValidateMetadataCols}
\title{ValidateMetadataCols}
\usage{
ValidateMetadataCols(seurat, cols)
}
\arguments{
\item{seurat}{A Seurat object}

\item{cols}{Character vector of metadata column names to check}
}
\value{
Invisibly returns NULL
}
\description{
Stop with a listing of any \code{cols} not present in a Seurat
  object's metadata. Intended for plot scripts to validate user-supplied
  column names up front, before the (slow) plotting loop.
}
\examples{

ValidateMetadataCols(pbmc_small, c("orig.ident", "groups"))

}
