% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiome_small}
\alias{multiome_small}
\title{Multiome Seurat object}
\format{
Contains two assays, RNA and ATAC. ATAC assay is a ChromatinAssay created with Signac
}
\usage{
multiome_small
}
\description{
Multiome Seurat object with 6 samples (20 cells each)
}
\keyword{datasets}
