% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveMetadata.R
\name{SaveMetadata}
\alias{SaveMetadata}
\title{SaveMetadata}
\usage{
SaveMetadata(seurat, filename, colname_for_rows = "cellid")
}
\arguments{
\item{seurat}{A Seurat object}

\item{filename}{A TSV filename}

\item{colname_for_rows}{The name used for the column name of the cell rownames}
}
\description{
Save the Seurat metadata to a TSV file.
}
\examples{
\dontrun{
SaveMetadata(pbmc_small, "pbmc_small_metadata.tsv")
}

}
