% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratInfo.R
\name{SeuratInfo}
\alias{SeuratInfo}
\title{SeuratInfo}
\usage{
SeuratInfo(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\description{
Show information about the Seurat object such as a table of the Idents and the first two rows of metadata. In addition, shows the available Reductions and Graphs. It shows a table of information about the Assays in the object and shows which Assay is the default.
}
\examples{
\dontrun{
SeuratInfo(seurat)
}
}
