% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignacInfo.R
\name{SignacInfo}
\alias{SignacInfo}
\title{SignacInfo}
\usage{
SignacInfo(seurat)
}
\arguments{
\item{seurat}{A Seurat object}
}
\description{
Show information about the Signac ChromatinAssays within a Seurat object
}
\examples{
\dontrun{
SignacInfo(pbmc_small)
}
}
