% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSample_scmapCluster.R
\name{TwoSample_scmapCluster}
\alias{TwoSample_scmapCluster}
\title{TwoSample_scmapCluster}
\usage{
TwoSample_scmapCluster(sce1, sce2, clustercols = "RNA_snn_res.0.8")
}
\arguments{
\item{sce1}{A SingleCellExperiment object}

\item{sce2}{Same as sce1. Must have same number of column names in 'colData'.}

\item{clustercols}{Metadata column(s) to use for clustering}
}
\value{
A list containing two lists ('1vs2' and '2vs1'). Each list contains
a set of scmapCluster() objects. Sankey plots for sce1 vs sce2 and Sankey
plots for sce2 vs sce1. Each plot will open in a browser.
}
\description{
This function performs a scmap cluster analysis on two Single Cell Experiment
objects.
}
\examples{
\dontrun{
TwoSample_scmapCluster(sce1, sce2)
}
}
\keyword{scmap}
