% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatePlotInputs.R
\name{ValidateReduction}
\alias{ValidateReduction}
\title{ValidateReduction}
\usage{
ValidateReduction(seurat, reduction)
}
\arguments{
\item{seurat}{A Seurat object}

\item{reduction}{Name of the reduction to check}
}
\value{
Invisibly returns NULL
}
\description{
Stop, listing the available reductions, if \code{reduction}
  is not present in a Seurat object.
}
\examples{

ValidateReduction(pbmc_small, "pca")

}
